实战开发04-多功能工具箱/requirements.txt
pyqt5
pyserial
pybluez2==0.37

# pip install -r .\requirements.txt